+++ src/components/menu-item/menu-item.test.ts
@@ -3,7 +3,7 @@ import sinon from 'sinon';
 import type SlMenuItem from './menu-item';
 
 describe('<sl-menu-item>', () => {
-  it('passes accessibility test', async () => {
+  it('should pass accessibility tests', async () => {
     const el = await fixture<SlMenuItem>(html`
       <sl-menu>
         <sl-menu-item>Item 1</sl-menu-item>
@@ -17,7 +17,7 @@ describe('<sl-menu-item>', () => {
     await expect(el).to.be.accessible();
   });
 
-  it('default properties', async () => {
+  it('should have the correct default properties', async () => {
     const el = await fixture<SlMenuItem>(html` <sl-menu-item>Test</sl-menu-item> `);
 
     expect(el.value).to.equal('');
@@ -25,7 +25,7 @@ describe('<sl-menu-item>', () => {
     expect(el.getAttribute('aria-disabled')).to.equal('false');
   });
 
-  it('changes aria attributes', async () => {
+  it('should render the correct aria attributes when disabled', async () => {
     const el = await fixture<SlMenuItem>(html` <sl-menu-item>Test</sl-menu-item> `);
 
     el.disabled = true;
@@ -33,12 +33,12 @@ describe('<sl-menu-item>', () => {
     expect(el.getAttribute('aria-disabled')).to.equal('true');
   });
 
-  it('get text label', async () => {
+  it('should return a text label when calling getTextLabel()', async () => {
     const el = await fixture<SlMenuItem>(html` <sl-menu-item>Test</sl-menu-item> `);
     expect(el.getTextLabel()).to.equal('Test');
   });
 
-  it('emits the slotchange event when the label changes', async () => {
+  it('should emit the slotchange event when the label changes', async () => {
     const el = await fixture<SlMenuItem>(html` <sl-menu-item>Text</sl-menu-item> `);
     const slotChangeHandler = sinon.spy();
 
@@ -47,5 +47,18 @@ describe('<sl-menu-item>', () => {
     await waitUntil(() => slotChangeHandler.calledOnce);
 
     expect(slotChangeHandler).to.have.been.calledOnce;
+  });
+
+  it('should render a hidden menu item when the inert attribute is used', async () => {
+    const menu = await fixture<SlMenuItem>(html`
+      <sl-menu>
+        <sl-menu-item inert>Item 1</sl-menu-item>
+        <sl-menu-item>Item 2</sl-menu-item>
+        <sl-menu-item>Item 3</sl-menu-item>
+      </sl-menu>
+    `);
+    const item1 = menu.querySelector('sl-menu-item')!;
+
+    expect(getComputedStyle(item1).display).to.equal('none');
   });
 });